can build

+++ token/src/token.rs
@@ -6,7 +6,7 @@ pub enum Token {
     //
     // Literal
     //
-    CharLiteral(i8),
+    CharLiteral(u32),
     // UCharLiteral(u8),
     // ShortLiteral(i16),
     // UShortLiteral(u16),
